Changes for supporting variable length tunnel encap metadata.
This support is enabled via a new nxm "tun_metadata". Multiple of these
nxms of variable sizes can appear in a flowmod.

Framework changes
=================

The nxms are learned dynamically and a hash-map is constructed that
maps <type,class> -> <length,offset> of the metadata. The map is
populated during flowmod via nxm_add, we allow a single length per
<type,class>. The length itself is not stored in the flow, hence
this limitation. Since the nxm field lookup assumes a fixed length,
the length needs to be adjusted before the lookup so we get the
correct nxm. For nxm of type MFF_TUN_METADATA we will also allow
duplicates.

In the first phase we do not reuse offsets from the map when flows
are removed. In the future this would be desirable.

The datapath that consumes this api (geneve in this case) needs to
provide a mapping function that converts netlink attr to the corresponding
metadata in the right order (after consulting the map, see
geneve_nlattr_to_tun_metadata)

TODO
====

1. dump flow miniflow expand consult the map and generate multiple nxms.
2. add ovs-appctl api to dump and delete the map.