Handle holes in tun_metadata.
When multiple flows use a subset of the created tun_metadata options
the tun_metadata within the miniflow can have holes filled with 0s.

We need to be able to figure out the length of the 0 filled memory. Since
tun_metadata can start with a 0, it cannot be done by counting 0s.

Hence add a new map to the tun_meta_table that maps an offset within
tun_metadata to length of the offset.

Modify ovs-ofctl dump-flows and appctl/trace to use the new api
find_tun_meta_ofs() to traverse the metadata.